Time: 11 ms (31.87%), Space: 8 MB (17.76%) - LeetHub

+++ 2240-number-of-ways-to-buy-pens-and-pencils/2240-number-of-ways-to-buy-pens-and-pencils.cpp
@@ -3,7 +3,7 @@ public:
     long long waysToBuyPensPencils(int total, int cost1, int cost2) {
         long long ans=0;
         for( int i=0;i*cost1<=total;i++){
-            int rem=total-i*cost1;
+            long long  rem=total-i*cost1;
             ans+=(rem/cost2)+1;
 
         }
